cleaning up the folder structure
