--- src/C/starpattern.c
@@ -1,30 +1,0 @@
-#include <stdio.h>
-int main(void)
-{
-   int i, j, rows;
-   printf("Enter the number of rows: ");
-   scanf("%d", &rows);
-   for (i = 1; i <= rows; ++i) 
-   {
-      	for (j = 1; j <= i; ++j)
-     	{
-         	    printf("* ");
-     	}
-        printf("\n");
-   }
-   return 0;
-}
-
-/*
-Half pyramid of *
-
-*
-* *
-* * *
-* * * *
-* * * * *
-
-
-
-
-*/
