+++ src/java/diamond.java
