+++ src/javascript/star.js
