+++ src/python/HollowSquare.py
@@ -1,22 +1,16 @@
-# Here 'n' is the number of rows and columns
+def hollowSquare(length, width):
+    for i in range(length):
+        if i == 0 or i == length - 1:
+            print("* " * width)
+        else:
+            #The length of the spacing could also be calculated, but this is a basic implementation
+            output = ""
+            for j in range(width):
+                if j == 0 or j == width - 1:
+                    output += "* "
+                else:
+                    output += "  "
+            print(output)
 
-def pattern(n):
-    print("* "*n,end="\n")
-    for a in range(n-2):
-        print("* "+"  "*(n-2)+"*")
-    print("* "*n,end="\n")
 
-pattern(5)
-
-
-'''
-
-this will return 
-
-* * * * *
-*       *
-*       *
-*       *
-* * * * *
-
-'''
+hollowSquare(5, 5)
