--- src/python/HollowSquarePattern.py
@@ -1,17 +1,0 @@
-for i in range(5):
-    if (i == 0 or i == 4):
-        for j in range(5):
-            print("*",end = " ")
-    else:
-        print("*       *",end = "")
-    print()
-
-
-"""
-* * * * * 
-*       *
-*       *
-*       *
-* * * * * 
-
-"""
