--- src/python/StarPattern.py
@@ -1,17 +1,0 @@
-
-row = int(input())
-for j in range(1,row):
-    for i in range(j):
-        print("*",end = " ")
-
-
-    print("")
-
-"""
-for 5 rows
-
-* 
-* * 
-* * * 
-* * * * 
-"""
