+++ src/python/leftArrow.py
