+++ src/python/square.py
@@ -1,11 +1,11 @@
 
-def square_pattern(n):
+def squarePattern(n):
   for i in range(n):
     for j in range(n):
       print("*",end=" ")
     print()
 
-square_pattern(5)
+squarePattern(5)
 
 '''
 python3 squarepattern.py
