--- src/python/squarePattern.py
@@ -1,17 +1,0 @@
-def squarePattern(n):
-    for i in range(n):
-       print("* " * n)
-
-
-n=int(input())
-squarePattern(n)
-
-
-
-/* 
-* * * * * 
-* * * * * 
-* * * * * 
-* * * * * 
-* * * * * 
-*/
